+++ sample/src/androidMain/kotlin/com/github/panpf/sketch/sample/ui/test/ShareElementTestFragment.kt
@@ -46,7 +46,7 @@ class ShareElementTestFragment : BaseBindingFragment<FragmentTestShareElementBin
             Log.i("ShareElementTest", "$id. loadImage")
             postponeEnterTransition(100, MILLISECONDS)
             loadImage(ResourceImages.jpeg.uri) {
-                registerListener(
+                addListener(
                     onSuccess = { _, _ ->
                         startPostponedEnterTransition()
                     },